+++ testExamples/ex3.correct.test.pas
@@ -1,0 +1,20 @@
+unit ex3;
+
+interface
+
+uses
+    Unit1
+  , Unit2
+  , Unit3
+  , Unit4
+  ;
+
+implementation
+uses
+    UnitA
+  , UnitB
+  , UnitC
+  , UnitD
+  ;
+
+end.