+++ testExamples/ex3.original.test.pas
@@ -1,0 +1,16 @@
+unit ex3;
+
+interface
+
+uses
+  Unit4, 
+  Unit2, 
+  Unit1, 
+  Unit3
+  ;
+
+implementation
+uses
+  UnitA, UnitC, UnitD, UnitB;
+
+end.