+++ testExamples/ex7.correct.test.pas
@@ -1,0 +1,18 @@
+unit Ex7.Original.Test;
+
+interface
+
+uses
+  System.Classes,
+  System.SysUtils,
+  Vcl.Controls,
+  Vcl.Forms;
+
+implementation
+
+uses
+  System.Math,
+  Vcl.Graphics,
+  Winapi.Windows;
+
+end.